midterm/requirements.txt
chainlit
python-dotenv
langchain
langchain_community
langchain_openai
langchain_qdrant
qdrant_client
